cart-service/requirements.txt
asgiref==3.11.1
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
